fix for faulty BGP session collisions
The code that maintains the in/out connection lists wasn't
being invoked in all cases where the connection was closed.

Moved the scattered connectionClosed() calls on
self.bgpPeering down into the common _closeConnection() code
so that they're always invoked.

Without this patch, in testing on lvs3002 with a live BGP
connection, if the connection is manually killed via
tcpkill, pybal would go into a recurring pattern of failing
reconnects, claiming a collision.  After this patch,
the first reconnect attempt succeeds and the daemon
carries on.

RT #7535
Change-Id: I98ecf9723c3f401ba23ffd63becf7b1eb3cf3329