runcommand: Twisted 13.x compatibility
As twisted.internet.process.Process has evolved significantly over the
years, the old ProcessGroupProcess implementation was now incompatible
with it. By exploiting the better structure of newer version of the base
class, we avoid reimplementing parts of the logic, and just add our
additional features (process groups support and timeouts) to the base
class.

Please note that this is /not/ compatible with twisted 11.x which we
have on Ubuntu Precise, so if we decide to merge this we should probably
create a precise branch of pybal where we can backport other, compatible changes.

Change-Id: I84dd674cd803c523bcdad1072108266778965aa8