Introduce ConfigurationObserver class
`ConfigurationObserver` is an interface. Its concrete implementations are
expected to handle a configuration URI scheme (like 'file://', 'http://',
'etcd://', etc.). The job of the ConfigurationObserver is to implement polling
of a server configuration resource for changes, fetching and parsing the
configuration, and notifying its associated Coordinator object, so it can act
on any configuration changes.

The patch includes one concrete subclass, 'FileConfigurationObserver', which
handles the 'file://' URI scheme, using inotify to watch the specified
configuration file for changes, and parsing it by following PyBal's existing
approach to pool configuration files, except replacing eval() with the safer
ast.literal_eval().

Change-Id: Icb42149e9e7395d4e7672b86ddcbbf7d78392292