Add instrumentation
Pybal has always been pretty private on its own internal state to
outside observers, making it a bit difficult to monitor or poll its
state.

We add a simple HTTP listener chain of monitoring endpoints describing
the actual state (pooled/depooled, up/down) of all servers (or a single
one) in a pool

Url space is left free for others to use for stats or other kind of
instrumentation in the future.

Bug: T102394
Change-Id: Ied88d44e63f91df3816df4b44f29bb0cd444cd05