Convert logging from print to twisted.python.log
Since moving to systemd, we decided it was a good idea to let journalctl
handle our logs. However, journald interacts badly with programs logging
to stdout as it imposes buffering, which we don't like.

Moreover, pybal logs are at the moment all-or-nothing, resulting in
sometimes precious yet verbose info to flood the pybal logs. What we
want is to be able to select (by log level or via syslog) the critical
information, what is informative, and what is most of the times
unnecessary and to get just what we need. We write our logs to stderr,
where they will be handled by journald directly and won't need any
additional intervention on our side.

As a consequence, we move to using twisted.python.log to log messages,
with some sugar on top that should make the transition to using
twisted.logger in newer twisted versions much easier.

Change-Id: I77ddd0c4f98373b21ea3dc0ba05cafaad757c029