IdleConnection: set keepalive
Enable the SO_KEEPALIVE socket option on the IdleConnection monitor's socket,
so that PyBal is able to gracefully handle the case when the peer terminates
the connection without letting us know.

Because the MemoryReactor used in PyBal (and Twisted)'s unit tests creates
`_FakeConnector` objects which do not have a transport property, the code
has to handle that case safely.

Bug: T113151
Change-Id: I99237b508426c0148671d5c4dfc916a89dfc89ce